+++ core/src/main/java/nl/nn/adapterframework/filesystem/FileSystemPipe.java
@@ -38,6 +38,7 @@ import nl.nn.adapterframework.stream.Message;
 import nl.nn.adapterframework.stream.MessageOutputStream;
 import nl.nn.adapterframework.stream.StreamingException;
 import nl.nn.adapterframework.stream.StreamingPipe;
+import nl.nn.adapterframework.util.SpringUtils;
 
 /**
  * Base class for Pipes that use a {@link IBasicFileSystem FileSystem}.
@@ -62,7 +63,9 @@ public abstract class FileSystemPipe<F, FS extends IBasicFileSystem<F>> extends 
 	@Override
 	public void configure() throws ConfigurationException {
 		super.configure();
-		getFileSystem().configure();
+		FS fileSystem = getFileSystem();
+		SpringUtils.autowireByName(getApplicationContext(), fileSystem);
+		fileSystem.configure();
 		try {
 			actor.configure(fileSystem, getParameterList(), this);
 		} catch (ConfigurationException e) {