+++ core/src/main/java/nl/nn/adapterframework/stream/json/JsonTee.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2021 WeAreFrank!
+   Copyright 2021, 2022 WeAreFrank!
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -75,6 +75,12 @@ public class JsonTee implements JsonEventHandler {
 	public void primitive(Object value) throws SAXException {
 		first.primitive(value);
 		second.primitive(value);
+	}
+
+	@Override
+	public void number(String value) throws SAXException {
+		first.number(value);
+		second.number(value);
 	}
 
 }