+++ core/src/test/java/nl/nn/adapterframework/xslt/XsltExceptionTest.java
@@ -41,7 +41,7 @@ public class XsltExceptionTest {
 			TransformerFilter transformer = tp.getTransformerFilter(threadConnector, filter);
 			
 			try {
-				try (SaxDocumentBuilder seb = new SaxDocumentBuilder("root", transformer)) {
+				try (SaxDocumentBuilder seb = new SaxDocumentBuilder("root", transformer, false)) {
 					seb.addElement("elem");
 					seb.addElement("error");
 					for(int i=0; i<tailCount; i++) {